blah

+++ src/main/java/cis579/ai/CardTracker.java
@@ -86,7 +86,7 @@ public class CardTracker {
 		}
 	}
 
-	public static double suggestionsMade() {
+	public static int suggestionsMade() {
 		return SUGGESTIONS_MADE;
 	}
